+4XP: Moving the code that builds the world out of the Renderer and into a World class. Simplifying the Simulation class's method signatures.
