+++ src/PlusForXP/renderer.ts
@@ -1,8 +1,6 @@
 import {
     GfxBlendFactor,
     GfxBlendMode,
-    GfxBufferFrequencyHint,
-    GfxBufferUsage,
     GfxCompareMode,
     GfxCullMode,
     GfxDevice,
@@ -10,9 +8,6 @@ import {
     GfxInputLayout,
     GfxRenderProgramDescriptor,
     GfxSampler,
-    GfxTexture,
-    GfxTextureDimension,
-    GfxTextureUsage,
     GfxVertexBufferFrequency,
 } from "../gfx/platform/GfxPlatform.js";
 import { TextureBase, TextureHolder } from "../TextureHolder.js";
@@ -24,54 +19,29 @@ import { fillMatrix4x4, fillVec4 } from "../gfx/helpers/UniformBufferHelpers.js"
 import { makeAttachmentClearDescriptor, makeBackbufferDescSimple, standardFullClearRenderPassDescriptor } from "../gfx/helpers/RenderGraphHelpers.js";
 import { GfxrAttachmentSlot } from "../gfx/render/GfxRenderGraph.js";
 import { colorNewFromRGBA } from "../Color.js";
-import { mat4, quat, vec3, vec4 } from "gl-matrix";
+import { mat4, vec3 } from "gl-matrix";
 import { GfxWrapMode, GfxTexFilterMode, GfxMipFilterMode } from "../gfx/platform/GfxPlatform.js";
-import { Camera, CameraController } from "../Camera.js";
+import { CameraController } from "../Camera.js";
 import { setAttachmentStateSimple } from "../gfx/helpers/GfxMegaStateDescriptorHelpers.js";
-import { bakeLights } from "./bake_lights.js";
-import { Material, Texture, SceneNode, Simulation, EnvironmentMap, ComputedEnvironmentMap, UnbakedMesh } from "./types.js";
+import { SceneNode, Simulation, WorldData } from "./types.js";
 import { SceneGfx, ViewerRenderInput } from "../viewer.js";
 import Plus4XPProgram from "./program.js";
-import { AnimationBuilder } from "./animation.js";
 import * as UI from "../ui.js";
-import { createSceneNode, createDataBuffer, updateNodeTransform, reparent, cloneTransform } from "./util.js";
-
-type Context = {
-    basePath: string;
-    scenes: Record<string, { scene: SCX.Scene; envID?: string }>;
-    textures: Texture[];
-    environmentMaps: Record<string, EnvironmentMap>;
-    cameras: [string, string][];
-    simulateFunc?: () => Simulation;
-};
+import { updateNodeTransform } from "./util.js";
+import { World } from "./world.js";
 
 export default class Renderer implements SceneGfx {
     private inputLayout: GfxInputLayout;
     private renderHelper: GfxRenderHelper;
     private renderInstListMain = new GfxRenderInstList();
     private program: GfxRenderProgramDescriptor;
-    private diffuseSampler: GfxSampler | null;
-    private envSampler: GfxSampler | null;
-    private fallbackMaterial: Material;
+    private diffuseSampler: GfxSampler;
+    private envSampler: GfxSampler;
 
-    private rootNode: SceneNode;
-    private sceneRot: vec3 = vec3.fromValues(-Math.PI / 2, 0, 0);
-    private inverseSceneRot: vec3 = vec3.negate(vec3.create(), this.sceneRot);
-    private materialsByName = new Map<string, Material>();
-    private sceneNodesByName = new Map<string, SceneNode>();
-    private renderableNodes: SceneNode[] = [];
-    private animatableNodes: SceneNode[] = [];
+    private world: World;
 
-    private camerasByName = new Map<string, SCX.Camera>();
-    private customCamera: Camera;
     private cameras: [string, string | null][];
     private activeCameraName: string | null = null;
-
-    private unbakedMeshes: UnbakedMesh[] = [];
-    private texturesByPath: Map<string, Texture>;
-    private environmentMapsByID = new Map<string, ComputedEnvironmentMap>();
-    private defaultTexture: GfxTexture;
-    private defaultEnvMap: ComputedEnvironmentMap;
 
     private simulation: Simulation | null;
     private cameraSelect: UI.SingleSelect;
@@ -81,29 +51,18 @@ export default class Renderer implements SceneGfx {
     private scratchWorldInverseTransposeMatrix = mat4.create();
 
     constructor(
-        private device: GfxDevice,
-        context: Context,
+        device: GfxDevice,
+        worldData: WorldData,
         public textureHolder: TextureHolder<TextureBase>,
     ) {
-        this.setupGraphics();
-
-        this.rootNode = createSceneNode({ name: "root" }, { rot: this.sceneRot });
-        for (const [name, { scene, envID }] of Object.entries(context.scenes)) {
-            this.buildScene(name, scene, envID);
-        }
-
-        this.customCamera = new Camera();
-        this.cameras = [["FreeCam", null], ...context.cameras];
-
-        this.bakeLights();
-        this.buildTextures(context.textures, context.environmentMaps);
-
-        this.simulation = context.simulateFunc?.() ?? null;
-        this.simulation?.setup?.(device, this.texturesByPath, this.materialsByName, this.sceneNodesByName);
+        this.setupGraphics(device);
+        this.world = new World(device, worldData);
+        this.cameras = [["FreeCam", null], ...worldData.cameras];
+        this.simulation = worldData.simulateFunc?.() ?? null;
+        this.simulation?.setup?.(device, this.world);
     }
 
-    private setupGraphics() {
-        const device = this.device;
+    private setupGraphics(device: GfxDevice) {
         setAttachmentStateSimple(
             { cullMode: GfxCullMode.Back },
             {
@@ -138,217 +97,6 @@ export default class Renderer implements SceneGfx {
         };
         this.diffuseSampler = device.createSampler(samplerDescriptor);
         this.envSampler = device.createSampler(samplerDescriptor);
-        this.fallbackMaterial = {
-            shader: {
-                name: "fallback",
-                id: -1,
-                ambient: vec3.create(),
-                diffuse: vec3.fromValues(1, 1, 1),
-                specular: vec3.fromValues(1, 1, 1),
-                opacity: 1,
-                luminance: 1,
-                blend: 0,
-            },
-            gfxTexture: null,
-        };
-    }
-
-    private buildScene(sceneName: string, scene: SCX.Scene, envID: string | undefined) {
-        const sceneRoot: SceneNode = createSceneNode({ name: sceneName + "_root", parent: this.rootNode });
-        this.sceneNodesByName.set(sceneRoot.name, sceneRoot);
-        this.rootNode.children.push(sceneRoot);
-        scene.lights.push({
-            type: SCX.LightType.Ambient,
-            name: "ambient",
-            color: scene.global.ambient,
-            intensity: 1,
-        });
-
-        for (const shader of scene.shaders) {
-            this.buildMaterial(sceneName, shader);
-        }
-
-        for (const camera of scene.cameras) {
-            this.buildCamera(sceneName, camera);
-        }
-
-        const nodes = new Map<string, SceneNode>();
-        for (const object of scene.objects) {
-            const node = this.buildObject(sceneName, scene, sceneRoot, object, envID);
-            nodes.set(node.name, node);
-        }
-        for (const sceneNode of nodes.values()) {
-            reparent(sceneNode, nodes.get(sceneNode.parentName ?? "") ?? sceneRoot);
-        }
-    }
-
-    private buildMaterial(sceneName: string, shader: SCX.Shader) {
-        const material = { shader, gfxTexture: null };
-        this.materialsByName.set(sceneName + shader.id, material);
-        return material;
-    }
-
-    private buildCamera(sceneName: string, camera: SCX.Camera) {
-        const cameraName = sceneName + camera.name;
-        const node: SceneNode = createSceneNode({ name: cameraName, loops: true }, { trans: camera.pos, rot: this.inverseSceneRot });
-        this.sceneNodesByName.set(cameraName, node);
-        reparent(node, this.rootNode);
-        this.camerasByName.set(cameraName, camera);
-        if (camera.animations !== undefined) {
-            node.animatedTransform = cloneTransform(node.transform);
-            node.animations = AnimationBuilder.build(node.animatedTransform!, camera.animations);
-            node.animates = node.animations.length > 0;
-            this.animatableNodes.push(node);
-        }
-        return node;
-    }
-
-    private buildObject(sceneName: string, scene: SCX.Scene, sceneRoot: SceneNode, object: SCX.Object, envID: string | undefined) {
-        const device = this.device;
-        const objectName = sceneName + object.name;
-        const node: SceneNode = createSceneNode({
-            name: objectName,
-            parentName: object.parent === null ? undefined : sceneName + object.parent,
-            transform: cloneTransform(object.transform),
-            loops: true,
-        });
-
-        for (const mesh of object.meshes ?? []) {
-            if (mesh.indices.length <= 0) {
-                continue;
-            }
-
-            const material = this.materialsByName.get(sceneName + mesh.shader) ?? this.fallbackMaterial;
-            if (material === this.fallbackMaterial) {
-                console.warn(`Missing shader ${mesh.shader} on mesh in ${object.name} of scene ${sceneName}. Falling back to default material.`);
-            }
-
-            const diffuseColorBuffer = device.createBuffer(mesh.vertexcount * 4, GfxBufferUsage.Vertex, GfxBufferFrequencyHint.Static);
-            device.uploadBufferData(diffuseColorBuffer, 0, new Uint8Array(new Float32Array(mesh.vertexcount * 4).fill(1).buffer));
-
-            const positionBuffer = createDataBuffer(device, GfxBufferUsage.Vertex, mesh.positions.buffer, mesh.dynamic);
-            const normalBuffer = createDataBuffer(device, GfxBufferUsage.Vertex, mesh.normals.buffer, mesh.dynamic);
-            const texcoordBuffer = createDataBuffer(device, GfxBufferUsage.Vertex, mesh.texCoords.buffer);
-            const vertexAttributes = [
-                { name: "position", ...(mesh.dynamic ? { data: mesh.positions } : null), buffer: positionBuffer, byteOffset: 0 },
-                { name: "normal", ...(mesh.dynamic ? { data: mesh.normals } : null), buffer: normalBuffer, byteOffset: 0 },
-                { name: "diffuseColor", buffer: diffuseColorBuffer, byteOffset: 0 },
-                { name: "texCoord", buffer: texcoordBuffer, byteOffset: 0 },
-            ];
-
-            const indexBuffer = createDataBuffer(device, GfxBufferUsage.Index, mesh.indices.buffer);
-            const indexBufferDescriptor = { buffer: indexBuffer, byteOffset: 0, ...(mesh.dynamic ? { data: mesh.indices } : null) };
-
-            this.unbakedMeshes.push({ node, mesh, shader: material.shader, diffuseColorBuffer, sceneName, lights: scene.lights });
-
-            node.meshes.push({
-                inputLayout: this.inputLayout,
-                vertexAttributes,
-                indexBufferDescriptor,
-                indexCount: mesh.indices.length,
-                material,
-                envID,
-            });
-        }
-
-        this.sceneNodesByName.set(objectName, node);
-        if (node.meshes.length > 0) {
-            this.renderableNodes.push(node);
-        }
-
-        if (object.animations !== undefined) {
-            node.animatedTransform = cloneTransform(node.transform);
-            node.animations = AnimationBuilder.build(node.animatedTransform!, object.animations);
-            node.animates = node.animations.length > 0;
-            this.animatableNodes.push(node);
-        }
-
-        return node;
-    }
-
-    private buildTextures(textures: Texture[], environmentMaps: Record<string, EnvironmentMap>) {
-        const device = this.device;
-
-        this.texturesByPath = new Map(textures.map((texture) => [texture.path, texture]));
-
-        for (const [envID, { texturePath, rotation, tint }] of Object.entries(environmentMaps)) {
-            const envTexture = this.texturesByPath.get(texturePath)!;
-            const texture = device.createTexture({
-                ...envTexture,
-                dimension: GfxTextureDimension.n2D,
-                pixelFormat: GfxFormat.U8_RGBA_NORM,
-                depthOrArrayLayers: 1,
-                numLevels: 1,
-                usage: GfxTextureUsage.Sampled,
-            });
-            device.uploadTextureData(texture, 0, [envTexture.rgba8]);
-            const matrix = mat4.fromQuat(mat4.create(), quat.fromEuler(quat.create(), ...rotation));
-            const computedTint = vec4.fromValues(...(tint ?? [1, 1, 1]), 1);
-            this.environmentMapsByID.set(envID, { texture, matrix, tint: computedTint });
-        }
-
-        this.defaultTexture = device.createTexture({
-            width: 1,
-            height: 1,
-            dimension: GfxTextureDimension.n2D,
-            pixelFormat: GfxFormat.U8_RGBA_NORM,
-            depthOrArrayLayers: 1,
-            numLevels: 1,
-            usage: GfxTextureUsage.Sampled,
-        });
-        this.defaultEnvMap = {
-            texture: this.defaultTexture,
-            matrix: mat4.create(),
-            tint: vec4.fromValues(1, 1, 1, 1),
-        };
-        device.uploadTextureData(this.defaultTexture, 0, [new Uint8Array([0xff, 0x00, 0xff, 0xff])]);
-
-        const requiredTextures = new Map<string, Texture>();
-        for (const material of this.materialsByName.values()) {
-            const texture = material.shader.texture === undefined ? null : (this.texturesByPath.get(material.shader.texture.replaceAll("\\", "/")) ?? null);
-
-            if (texture === null) {
-                continue;
-            }
-
-            const texturePath = texture.path;
-            if (!requiredTextures.has(texturePath)) {
-                const gfxTexture = device.createTexture({
-                    ...texture,
-                    dimension: GfxTextureDimension.n2D,
-                    pixelFormat: GfxFormat.U8_RGBA_NORM,
-                    depthOrArrayLayers: 1,
-                    numLevels: 1,
-                    usage: GfxTextureUsage.Sampled,
-                });
-                device.uploadTextureData(gfxTexture, 0, [texture.rgba8]);
-                texture.gfxTexture = gfxTexture;
-                requiredTextures.set(texturePath, texture);
-            }
-            material.gfxTexture = requiredTextures.get(texture?.path)?.gfxTexture ?? null;
-        }
-    }
-
-    private bakeLights() {
-        updateNodeTransform(this.rootNode, false, null, false);
-        const transformedLightsBySceneName: Map<string, SCX.Light[]> = new Map();
-        const rootTransform = this.rootNode.worldTransform;
-        for (const { node, shader, mesh, diffuseColorBuffer, sceneName, lights } of this.unbakedMeshes) {
-            if (!transformedLightsBySceneName.has(sceneName)) {
-                transformedLightsBySceneName.set(
-                    sceneName,
-                    lights.map(
-                        (light: SCX.Light): SCX.Light => ({
-                            ...light,
-                            pos: light.pos === undefined ? undefined : vec3.transformMat4(vec3.create(), light.pos, rootTransform),
-                            dir: light.dir === undefined ? undefined : vec3.transformMat4(vec3.create(), light.dir, rootTransform),
-                        }),
-                    ),
-                );
-            }
-            const diffuseColors = bakeLights(mesh, shader, node.worldTransform, transformedLightsBySceneName.get(sceneName)!);
-            this.device.uploadBufferData(diffuseColorBuffer, 0, new Uint8Array(diffuseColors.buffer));
-        }
     }
 
     public adjustCameraController(c: CameraController) {
@@ -383,8 +131,8 @@ export default class Renderer implements SceneGfx {
         this.animating = deltaTime > 0;
 
         if (this.animating) {
-            this.simulation?.update?.(viewerInput, this.sceneNodesByName, device);
-            for (const node of this.animatableNodes) {
+            this.simulation?.update?.(device, viewerInput);
+            for (const node of this.world.animatableNodes) {
                 if (!node.animates) {
                     continue;
                 }
@@ -402,12 +150,12 @@ export default class Renderer implements SceneGfx {
         const mainColorTargetID = builder.createRenderTargetID(mainColorDesc, "Main Color");
         const mainDepthTargetID = builder.createRenderTargetID(mainDepthDesc, "Main Depth");
 
-        updateNodeTransform(this.rootNode, false, null, this.animating);
+        updateNodeTransform(this.world.rootNode, false, null, this.animating);
 
         if (this.animating) {
             const cameraWorldPos = mat4.getTranslation(
                 vec3.create(),
-                this.activeCameraName !== null ? this.sceneNodesByName.get(this.activeCameraName)!.worldTransform : viewerInput.camera.worldMatrix,
+                this.activeCameraName !== null ? this.world.sceneNodesByName.get(this.activeCameraName)!.worldTransform : viewerInput.camera.worldMatrix,
             );
             this.simulation?.render(this.renderHelper, builder, cameraWorldPos);
         }
@@ -444,14 +192,14 @@ export default class Renderer implements SceneGfx {
             this.lastViewerCameraMatrix ??= [...viewerInput.camera.worldMatrix].join("_");
 
             if (this.activeCameraName !== null) {
-                const camera: SCX.Camera = this.camerasByName.get(this.activeCameraName)!;
-                const cameraNode: SceneNode = this.sceneNodesByName.get(this.activeCameraName)!;
+                const camera: SCX.Camera = this.world.camerasByName.get(this.activeCameraName)!;
+                const cameraNode: SceneNode = this.world.sceneNodesByName.get(this.activeCameraName)!;
 
-                this.customCamera.clipSpaceNearZ = viewerInput.camera.clipSpaceNearZ;
-                this.customCamera.setPerspective(camera.fov, viewerInput.camera.aspect, camera.nearclip, camera.farclip);
+                this.world.customCamera.clipSpaceNearZ = viewerInput.camera.clipSpaceNearZ;
+                this.world.customCamera.setPerspective(camera.fov, viewerInput.camera.aspect, camera.nearclip, camera.farclip);
 
                 const cameraWorldPos = mat4.getTranslation(vec3.create(), cameraNode.worldTransform);
-                const targetWorldPos = vec3.transformMat4(vec3.create(), camera.targetpos, this.rootNode.worldTransform);
+                const targetWorldPos = vec3.transformMat4(vec3.create(), camera.targetpos, this.world.rootNode.worldTransform);
                 const relativePos = vec3.sub(vec3.create(), targetWorldPos, cameraWorldPos);
                 mat4.fromTranslation(this.scratchViewMatrix, cameraWorldPos);
                 mat4.rotateY(this.scratchViewMatrix, this.scratchViewMatrix, -Math.PI / 2 - Math.atan2(relativePos[2], relativePos[0]));
@@ -463,7 +211,7 @@ export default class Renderer implements SceneGfx {
                     viewerInput.camera.worldMatrixUpdated();
                     cameraOffset += fillMatrix4x4(cameraBuffer, cameraOffset, viewerInput.camera.projectionMatrix);
                 } else {
-                    cameraOffset += fillMatrix4x4(cameraBuffer, cameraOffset, this.customCamera.projectionMatrix);
+                    cameraOffset += fillMatrix4x4(cameraBuffer, cameraOffset, this.world.customCamera.projectionMatrix);
                 }
                 mat4.invert(this.scratchViewMatrix, this.scratchViewMatrix);
                 cameraOffset += fillMatrix4x4(cameraBuffer, cameraOffset, this.scratchViewMatrix);
@@ -491,13 +239,13 @@ export default class Renderer implements SceneGfx {
             depthWrite: !ghosts,
         });
 
-        for (const node of this.renderableNodes) {
+        for (const node of this.world.renderableNodes) {
             if (!node.worldVisible || node.isGhost !== ghosts) {
                 continue;
             }
 
             for (const mesh of node.meshes!) {
-                const envMap = (mesh.envID === undefined ? null : this.environmentMapsByID.get(mesh.envID)) ?? this.defaultEnvMap;
+                const envMap = (mesh.envID === undefined ? null : this.world.environmentMapsByID.get(mesh.envID)) ?? this.world.defaultEnvMap;
                 const renderInst = renderInstManager.newRenderInst();
                 updateObjectParams: {
                     let objectOffset = renderInst.allocateUniformBuffer(Plus4XPProgram.ub_ObjectParams, 16 * 3 + 4 + 4 /*Mat4x3 * 3 + vec4 * 2*/);
@@ -519,7 +267,7 @@ export default class Renderer implements SceneGfx {
 
                 renderInst.setSamplerBindingsFromTextureMappings([
                     {
-                        gfxTexture: mesh.material.gfxTexture ?? this.defaultTexture,
+                        gfxTexture: mesh.material.gfxTexture ?? this.world.defaultTexture,
                         gfxSampler: this.diffuseSampler,
                         lateBinding: null,
                     },
@@ -530,7 +278,7 @@ export default class Renderer implements SceneGfx {
                     },
                 ]);
 
-                renderInst.setVertexInput(mesh.inputLayout, mesh.vertexAttributes, mesh.indexBufferDescriptor);
+                renderInst.setVertexInput(this.inputLayout, mesh.vertexAttributes, mesh.indexBufferDescriptor);
 
                 renderInst.setDrawCount(mesh.indexCount);
                 renderInstManager.submitRenderInst(renderInst);
@@ -543,37 +291,10 @@ export default class Renderer implements SceneGfx {
     destroy(device: GfxDevice): void {
         this.simulation?.destroy(device);
         this.simulation = null;
+        this.world.destroy(device);
         this.textureHolder.destroy(device);
         this.renderHelper.destroy();
-        for (const material of this.materialsByName.values()) {
-            if (material.gfxTexture !== null) {
-                device.destroyTexture(material.gfxTexture);
-                material.gfxTexture = null;
-            }
-        }
-        this.materialsByName.clear();
-        for (const node of this.sceneNodesByName.values()) {
-            if (node.meshes === null || node.meshes.length === 0) {
-                continue;
-            }
-            for (const { buffer } of node.meshes[0].vertexAttributes) {
-                device.destroyBuffer(buffer);
-            }
-            for (const mesh of node.meshes) {
-                device.destroyBuffer(mesh.indexBufferDescriptor.buffer);
-            }
-        }
-        if (this.diffuseSampler !== null) {
-            device.destroySampler(this.diffuseSampler);
-        }
-        this.diffuseSampler = null;
-        if (this.envSampler !== null) {
-            device.destroySampler(this.envSampler);
-        }
-        this.envSampler = null;
-        this.sceneNodesByName.clear();
-        this.renderableNodes.length = 0;
-        this.camerasByName.clear();
-        this.rootNode.children = [];
+        device.destroySampler(this.diffuseSampler);
+        device.destroySampler(this.envSampler);
     }
 }
